Basic grid layout + content
This uses grid layout to implement a card-ish design that intentionally
puts some flexible space between elements (for example, the footer is
pushed to the bottom).

Sections have backgrounds that will give us some opacity over the
background.

The link suggestions are presented in a table. The idea is that
javascript will augment this table.